Fixed a bug in cerioli2010.irmcd.test pointed out by Yu Zhang: the roles of alpha and gamma were switched, so that gamma was being used for the FSRMCD test and alpha was being used for the IRMCD. Fixed that and verified correctness via simulation. (Note that original code passed the simulation tests too; the bug only seems to arise in larger samples, for which IRMCD and FSRMCD are not recommended.)
Also expanded simulation tests to include dimensions up to 20 and
n up to 6400.

Changed the name of the error rate in cerioli2010.irmcd.test from
'signif.alpha' to 'signif.gamma' to make it more clear that the IRMCD
function takes the Type I error rate for the intersection hypothesis
('gamma' in Cerioli (2010)) as input. Edited the documentation to make
this more clear.

Still to do:

1) Add option to use original Hardin-Rocke distribution in FSRMCD
(instead of Green-Martin modification).
2) Add arguments to allow some of the modifications mentioned in
Cerioli (2010), page 151.
2) Review documentation for improvements, particularly in examples.
Should add the testing code as a "don't run" example.
3) Add banknotes example.
4) Check under latest R and submit to CRAN.